Index games in batches
This involves paginating at the query level as it's not really practical
to keep pulling rows out a statement on a pooled connection (or so it
would seem). So now 100 games are pulled out at a time up until all
games are exhausted then we go into a sleeping/polling loop.